move to Decimal - also duplicate the memory consumption

+++ src/parser/gex/constraint.rs
@@ -1,3 +1,5 @@
+use rust_decimal::Decimal;
+
 /*
  * Used for constrained ranges that are too large to be pre-calculated
  * or for ranges with constraints that depend on non-static values, therefore
@@ -5,6 +7,6 @@
  */
 #[derive(Debug, Clone)]
 pub enum Constraint {
-    MultipleOf(f64), // Expects the Lowest Common Multiple of all the numbers. No expressions allowed here
-    NotMultipleOf(Vec<f64>), // Expects all the blacklisted numbers
+    MultipleOf(Decimal), // Expects the Lowest Common Multiple of all the numbers. No expressions allowed here
+    NotMultipleOf(Vec<Decimal>), // Expects all the blacklisted numbers
 }
